requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.110.0
feedparser==6.0.11
Flask==3.0.2
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.12.0
orjson==3.9.15
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
waitress==3.0.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
